LayoutTests/css3/scroll-snap/scroll-snap-iframe-rtl.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }

    var succeeded = false;
    var successfullyParsed = false;

    window.onload = function() {
        shouldBe("succeeded", "true");

        isSuccessfullyParsed();
    };
    </script>
</head>
<body>
    <p id="description"></p>
    <iframe id="frame" src="resources/iframe-content-rtl.html" style="width: 100vw; height: 100vh"></iframe>
    <script>
        description("Verify that the iframe propery parses snap-point settings.");
    </script>
</body>
</html>
